PI-24 - API Profile
Alterado type do campo id para Long

+++ backend/src/main/java/br/com/house/digital/projetointegrador/dto/ProfileNewDTO.java
@@ -44,7 +44,7 @@ public class ProfileNewDTO implements Serializable {
 	private List<Company> companys = new ArrayList<>();
 
 	@NotNull(message = "Mandatory Filling")	
-	private Integer userId;
+	private Long userId;
 
 	public ProfileNewDTO() {
